+++ src/main/java/com/app/estore/response/VendorProductDto.java
@@ -1,16 +1,13 @@
 package com.app.estore.response;
 
-import com.app.estore.entity.Vendor;
 import com.app.estore.utility.ProductCategory;
-import jakarta.persistence.*;
 import lombok.Data;
-import org.hibernate.annotations.CreationTimestamp;
 
 import java.time.LocalDateTime;
 
 @Data
 public class VendorProductDto {
-    private Integer productId;
+    private Long productId;
     private String title;
     private String description;
     private String weight;
